+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## In development
+### Compatibility
+- Smart Doors now uses the libwrapper module and as a result is now compatible with the module "FoundryVTT Arms Reach"
+
 ## 1.2.5
 ### New features
 - Synchronized doors can now be configured to synchronize their secret door status as well